+++ mlxtend/regressor/tests/test_stacking_regression.py
